(Azure CXP) Updated Regex
Updated Regular expression to make sure that the absolute path expression defined in the if statement works. 

Changed the following at line 105 
`	$filePath = if($FilePath -match "^[a-zA-Z]:\*") {$FilePath} else {"$($PWD.path)\$($FilePath)"}
`
to 
`     $filePath = if($FilePath -match "^[a-zA-Z]:\\.*") {$FilePath} else {"$($PWD.path)\$($FilePath)"}
`

In the statement absolute path does not work and only relative path works as complained by the customer on the Github issue MicrosoftDocs/azure-docs#51943 . If we change the regex in the script as above , it will be fixed. Hence requesting an update in the script . 

Thank you.